docker build fix
