+++ go/launcher/client/client-otlp.go
@@ -3,7 +3,8 @@
 //
 // usage:
 //	 export OTEL_LOG_LEVEL=debug
-//   go run client-collector.go
+//   export LS_ACCESS_TOKEN=<YOUR_LS_ACCESS_TOKEN>
+//   go run client-direct.go
 
 package main
 
@@ -28,17 +29,18 @@ var (
 	serviceName    = os.Getenv("LS_SERVICE_NAME")
 	serviceVersion = os.Getenv("LS_SERVICE_VERSION")
 	endpoint       = os.Getenv("LS_SATELLITE_URL")
+	lsToken        = os.Getenv("LS_ACCESS_TOKEN")
 	targetURL      = os.Getenv("DESTINATION_URL")
 )
 
 func newLauncher() launcher.Launcher {
 	if len(endpoint) == 0 {
-		endpoint = "localhost:4317" // Collector endpoint
+		endpoint = "ingest.lightstep.com:443"
 		log.Printf("Using default LS endpoint %s", endpoint)
 	}
 
 	if len(serviceName) == 0 {
-		serviceName = "test-go-client-launcher-collector"
+		serviceName = "test-go-client-launcher-direct"
 		log.Printf("Using default service name %s", serviceName)
 	}
 
@@ -47,13 +49,16 @@ func newLauncher() launcher.Launcher {
 		log.Printf("Using default service version %s", serviceVersion)
 	}
 
+	if len(lsToken) == 0 {
+		log.Fatalf("Lightstep token missing. Please set environment variable LS_ENVIRONMENT")
+	}
+
 	otelLauncher := launcher.ConfigureOpentelemetry(
 		launcher.WithServiceName(serviceName),
 		launcher.WithServiceVersion(serviceVersion),
-		launcher.WithSpanExporterInsecure(true), // Use for Collector
+		launcher.WithAccessToken(lsToken),
 		launcher.WithSpanExporterEndpoint(endpoint),
 		launcher.WithMetricExporterEndpoint(endpoint),
-		launcher.WithMetricExporterInsecure(true), // Use for Collector
 		launcher.WithPropagators([]string{"tracecontext", "baggage"}),
 		launcher.WithResourceAttributes(map[string]string{
 			string(semconv.ContainerNameKey): "my-container-name",
