+++ go/launcher/server/server.go
@@ -2,12 +2,12 @@
 // example code to test lightstep/opentelemetry-exporter-go
 //
 // usage:
+//	 export OTEL_LOG_LEVEL=debug
 //   go run server.go
 
 package main
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"math/rand"
@@ -16,15 +16,11 @@ import (
 	"strings"
 	"time"
 
+	"github.com/lightstep/otel-launcher-go/launcher"
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
-	"go.opentelemetry.io/otel/propagation"
-	"go.opentelemetry.io/otel/sdk/resource"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	semconv "go.opentelemetry.io/otel/semconv/v1.10.0"
 	"go.opentelemetry.io/otel/trace"
 )
@@ -34,8 +30,6 @@ var (
 	serviceName    = os.Getenv("LS_SERVICE_NAME")
 	serviceVersion = os.Getenv("LS_SERVICE_VERSION")
 	endpoint       = os.Getenv("LS_SATELLITE_URL")
-	lsToken        = os.Getenv("LS_ACCESS_TOKEN")
-	lsEnvironment  = os.Getenv("LS_ENVIRONMENT")
 )
 
 const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
@@ -47,32 +41,14 @@ const (
 
 var src = rand.NewSource(time.Now().UnixNano())
 
-func newExporter(ctx context.Context) (*otlptrace.Exporter, error) {
-
+func newLauncher() launcher.Launcher {
 	if len(endpoint) == 0 {
-		endpoint = "ingest.lightstep.com:443"
+		endpoint = "localhost:4317" // Collector endpoint
 		log.Printf("Using default LS endpoint %s", endpoint)
 	}
 
-	if len(lsToken) == 0 {
-		log.Fatalf("Lightstep token missing. Please set environment variable LS_ENVIRONMENT")
-	}
-
-	var headers = map[string]string{
-		"lightstep-access-token": lsToken,
-	}
-
-	client := otlptracegrpc.NewClient(
-		otlptracegrpc.WithHeaders(headers),
-		otlptracegrpc.WithEndpoint(endpoint),
-	)
-	return otlptrace.New(ctx, client)
-}
-
-func newTraceProvider(exp *otlptrace.Exporter) *sdktrace.TracerProvider {
-
 	if len(serviceName) == 0 {
-		serviceName = "test-go-server-grpc"
+		serviceName = "test-go-server-launcher-collector"
 		log.Printf("Using default service name %s", serviceName)
 	}
 
@@ -81,30 +57,20 @@ func newTraceProvider(exp *otlptrace.Exporter) *sdktrace.TracerProvider {
 		log.Printf("Using default service version %s", serviceVersion)
 	}
 
-	if len(lsEnvironment) == 0 {
-		lsEnvironment = "dev"
-		log.Printf("Using default environment %s", lsEnvironment)
-	}
-
-	resource, rErr :=
-		resource.Merge(
-			resource.Default(),
-			resource.NewWithAttributes(
-				semconv.SchemaURL,
-				semconv.ServiceNameKey.String(serviceName),
-				semconv.ServiceVersionKey.String(serviceVersion),
-				attribute.String("environment", lsEnvironment),
-			),
-		)
-
-	if rErr != nil {
-		panic(rErr)
-	}
-
-	return sdktrace.NewTracerProvider(
-		sdktrace.WithBatcher(exp),
-		sdktrace.WithResource(resource),
+	otelLauncher := launcher.ConfigureOpentelemetry(
+		launcher.WithServiceName(serviceName),
+		launcher.WithServiceVersion(serviceVersion),
+		launcher.WithSpanExporterInsecure(true), // Use for Collector
+		launcher.WithSpanExporterEndpoint(endpoint),
+		launcher.WithMetricExporterEndpoint(endpoint),
+		launcher.WithMetricExporterInsecure(true), // Use for Collector
+		launcher.WithPropagators([]string{"tracecontext", "baggage"}),
+		launcher.WithResourceAttributes(map[string]string{
+			string(semconv.ContainerNameKey): "my-container-name",
+		}),
 	)
+
+	return otelLauncher
 }
 
 func randString(n int) string {
@@ -143,7 +109,6 @@ func handlePing(w http.ResponseWriter, r *http.Request) {
 
 	pingResult := randString(length)
 	span.SetAttributes(
-		// attribute.String("result", pingResult),
 		attribute.String("library.language", "go"),
 		attribute.String("library.version", "v1.7.0"),
 	)
@@ -158,31 +123,15 @@ func handlePing(w http.ResponseWriter, r *http.Request) {
 }
 
 func main() {
-	ctx := context.Background()
+	otelLauncher := newLauncher()
+	defer otelLauncher.Shutdown()
 
-	exp, err := newExporter(ctx)
-	if err != nil {
-		log.Fatalf("failed to initialize exporter: %v", err)
-	}
-
-	tp := newTraceProvider(exp)
-	defer func() { _ = tp.Shutdown(ctx) }()
-
-	otel.SetTracerProvider(tp)
-
-	otel.SetTextMapPropagator(
-		propagation.NewCompositeTextMapPropagator(
-			propagation.TraceContext{},
-			propagation.Baggage{},
-		),
-	)
-
-	tracer = tp.Tracer(serviceName, trace.WithInstrumentationVersion(serviceVersion))
+	tracer = otel.Tracer(serviceName)
 
 	wrapHandler()
 
 	fmt.Printf("Starting server on http://localhost:8081\n")
-	err = http.ListenAndServe(":8081", nil)
+	err := http.ListenAndServe(":8081", nil)
 	if err != nil {
 		log.Fatal(err)
 	}
