+++ go/opentelemetry/otlp/server/server.go
@@ -2,21 +2,25 @@
 // example code to test lightstep/opentelemetry-exporter-go
 //
 // usage:
-//   go run client.go
+//   export LS_ACCESS_TOKEN=<YOUR_LS_ACCESS_TOKEN>
+//   go run server.go
 
 package main
 
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"log"
+	"math/rand"
+	"net/http"
 	"os"
+	"strings"
 	"time"
 
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
 	"go.opentelemetry.io/otel/propagation"
@@ -33,8 +37,16 @@ var (
 	endpoint       = os.Getenv("LS_SATELLITE_URL")
 	lsToken        = os.Getenv("LS_ACCESS_TOKEN")
 	lsEnvironment  = os.Getenv("LS_ENVIRONMENT")
-	targetURL      = os.Getenv("DESTINATION_URL")
 )
+
+const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+const (
+	letterIdxBits = 6                    // 6 bits to represent a letter index
+	letterIdxMask = 1<<letterIdxBits - 1 // All 1-bits, as many as letterIdxBits
+	letterIdxMax  = 63 / letterIdxBits   // # of letter indices fitting in 63 bits
+)
+
+var src = rand.NewSource(time.Now().UnixNano())
 
 func newExporter(ctx context.Context) (*otlptrace.Exporter, error) {
 
@@ -61,7 +73,7 @@ func newExporter(ctx context.Context) (*otlptrace.Exporter, error) {
 func newTraceProvider(exp *otlptrace.Exporter) *sdktrace.TracerProvider {
 
 	if len(serviceName) == 0 {
-		serviceName = "test-go-client-grpc"
+		serviceName = "test-go-server-grpc"
 		log.Printf("Using default service name %s", serviceName)
 	}
 
@@ -96,31 +108,54 @@ func newTraceProvider(exp *otlptrace.Exporter) *sdktrace.TracerProvider {
 	)
 }
 
-func makeRequest(ctx context.Context) {
-	ctx, span := tracer.Start(ctx, "makeRequest")
+func randString(n int) string {
+	sb := strings.Builder{}
+	sb.Grow(n)
+	for i, cache, remain := n-1, src.Int63(), letterIdxMax; i >= 0; {
+		if remain == 0 {
+			cache, remain = src.Int63(), letterIdxMax
+		}
+		if idx := int(cache & letterIdxMask); idx < len(letterBytes) {
+			sb.WriteByte(letterBytes[idx])
+			i--
+		}
+		cache >>= letterIdxBits
+		remain--
+	}
+
+	return sb.String()
+}
+
+// Wrap the handleRollDice so that telemetry data
+// can be automatically generated for it
+func wrapHandler() {
+	handler := http.HandlerFunc(handlePing)
+	wrappedHandler := otelhttp.NewHandler(handler, "pingHandler")
+	http.Handle("/ping", wrappedHandler)
+}
+
+func handlePing(w http.ResponseWriter, r *http.Request) {
+	operationName := "ping"
+	_, span := tracer.Start(r.Context(), operationName)
 	defer span.End()
 
-	if len(targetURL) == 0 {
-		targetURL = "http://localhost:8081/ping"
-		log.Printf("Using default targetURL %s", targetURL)
-	}
+	length := rand.Intn(1024)
+	log.Printf("%s %s %s", r.Method, r.URL.Path, r.Proto)
 
-	span.AddEvent("Did some cool stuff")
-	res, err := otelhttp.Get(ctx, targetURL)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	defer res.Body.Close()
-	body, _ := ioutil.ReadAll(res.Body)
-	fmt.Printf("Body : %s", body)
-	fmt.Printf("Request to %s, got %d bytes\n", targetURL, res.ContentLength)
-
+	pingResult := randString(length)
 	span.SetAttributes(
-		attribute.String("response", string(body)),
+		// attribute.String("result", pingResult),
+		attribute.String("library.language", "go"),
+		attribute.String("library.version", "v1.7.0"),
 	)
 
-	span.AddEvent("Cancelled wait due to external signal", trace.WithAttributes(attribute.Int("pid", 4328), attribute.String("signal", "SIGHUP")))
+	// setting span as successful
+	span.SetStatus(codes.Ok, "Success")
+
+	// setting span event
+	span.AddEvent(fmt.Sprint(r.Header))
+
+	fmt.Fprintf(w, pingResult)
 }
 
 func main() {
@@ -145,9 +180,11 @@ func main() {
 
 	tracer = tp.Tracer(serviceName, trace.WithInstrumentationVersion(serviceVersion))
 
-	for {
-		makeRequest(ctx)
-		time.Sleep(1 * time.Second)
-	}
+	wrapHandler()
 
+	fmt.Printf("Starting server on http://localhost:8081\n")
+	err = http.ListenAndServe(":8081", nil)
+	if err != nil {
+		log.Fatal(err)
+	}
 }
